server/requirements.txt
beautifulsoup4==4.11.1
celery==5.2.7
Django==3.2.7
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-filter==22.1
django-grappelli==3.0.3
django-import-export==2.8.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
flower==1.1.0
redis==4.3.4
requests==2.28.1
